+++ Example/Sources/Views/CameraInputBarAccessoryView.swift
@@ -42,7 +42,7 @@ class CameraInputBarAccessoryView: InputBarAccessoryView {
   }()
 
   func configure() {
-    let camera = makeButton(named: "ic_camera")
+    let camera = makeCameraButton()
     camera.tintColor = .darkGray
     camera.onTouchUpInside { [weak self] _ in
       self?.showImagePickerControllerActionSheet()
@@ -64,7 +64,7 @@ class CameraInputBarAccessoryView: InputBarAccessoryView {
 
   // MARK: Private
 
-  private func makeButton(named _: String) -> InputBarButtonItem {
+  private func makeCameraButton() -> InputBarButtonItem {
     InputBarButtonItem()
       .configure {
         $0.spacing = .fixed(10)
@@ -74,8 +74,6 @@ class CameraInputBarAccessoryView: InputBarAccessoryView {
         $0.tintColor = .systemBlue
       }.onDeselected {
         $0.tintColor = UIColor.lightGray
-      }.onTouchUpInside { _ in
-        print("Item Tapped")
       }
   }
 }
@@ -118,12 +116,10 @@ extension CameraInputBarAccessoryView: UIImagePickerControllerDelegate, UINaviga
     didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey: Any])
   {
     if let editedImage = info[UIImagePickerController.InfoKey.editedImage] as? UIImage {
-      // self.sendImageMessage(photo: editedImage)
       inputPlugins.forEach { _ = $0.handleInput(of: editedImage) }
     }
     else if let originImage = info[UIImagePickerController.InfoKey.originalImage] as? UIImage {
       inputPlugins.forEach { _ = $0.handleInput(of: originImage) }
-      // self.sendImageMessage(photo: originImage)
     }
     getRootViewController()?.dismiss(animated: true, completion: nil)
     inputAccessoryView?.isHidden = false
@@ -167,7 +163,6 @@ extension CameraInputBarAccessoryView: AttachmentManagerDelegate {
   // MARK: - AttachmentManagerDelegate Helper
 
   func setAttachmentManager(active: Bool) {
-    let topStackView = topStackView
     if active, !topStackView.arrangedSubviews.contains(attachmentManager.attachmentView) {
       topStackView.insertArrangedSubview(attachmentManager.attachmentView, at: topStackView.arrangedSubviews.count)
       topStackView.layoutIfNeeded()